use int64 for all bitwise operations

+++ test/ulid_test.dart
@@ -9,7 +9,7 @@ import 'utils.dart';
 void main() {
   test('invalid timestamp', () {
     expect(
-      () => requireTimestamp(0x0001000000000000),
+      () => requireTimestamp(Int64(0x0001000000000000)),
       throwsArgumentError,
     );
   });
